Use int for output buffer size
size_t is 8 bytes in 64-bit mode, but XMLSERVICE expects an int. By
passing a pointer to it, XMLSERVICE will only see the top 4 bytes which
would always be 0. This has the effect of producing no output, even
though XMLSERVICE handled the call.

+++ utils/xmlservice-cli.c
@@ -121,7 +121,7 @@ int main(int argc, char **argv)
         }
     }
 
-    size_t xmlout_len = (16 * 1024 * 1024);
+    int xmlout_len = (16 * 1024 * 1024);
     char* xmlout = (char*) malloc(xmlout_len);
 
     if(!xmlout) {
